refactor(ops): share the backstage ingest job poll helper
