+++ ops/backstage-lore-download-e2e.ts
@@ -38,10 +38,8 @@ import type {
   PublishBackstagePackageResult,
   PublishBackstageReleaseConfig,
 } from './publish-backstage-package';
+import { pollBackstageIngestJob } from './lib/backstageIngestPoll';
 
-const BACKSTAGE_INGEST_POLL_INTERVAL_MS = 1000;
-const BACKSTAGE_INGEST_POLL_REQUEST_TIMEOUT_MS = 15_000;
-const BACKSTAGE_INGEST_POLL_TIMEOUT_MS = 20 * 60 * 1000;
 const BACKSTAGE_TUS_CHUNK_SIZE = 64 * 1024 * 1024;
 const REPO_TOKEN_HEADER = 'X-YUCP-Repo-Token';
 const REPOSITORY_PATH = '/v1/backstage/repos/index.json';
@@ -68,11 +66,6 @@ type UploadAuthorization = {
   uploadMetadataKey: string;
   maxByteSize: number;
 };
-
-type BackstageIngestJobResponse =
-  | { state: 'processing' }
-  | { state: 'completed'; result: string }
-  | { state: 'failed'; reason: string };
 
 type RepositoryManifest = {
   headers: Record<string, string>;
@@ -330,45 +323,6 @@ async function authorizeUpload(
   return { tusEndpoint, uploadToken, uploadMetadataKey, maxByteSize: authorization.maxByteSize };
 }
 
-async function pollBackstageIngestJob(jobUrl: string, uploadToken: string): Promise<string> {
-  const deadline = Date.now() + BACKSTAGE_INGEST_POLL_TIMEOUT_MS;
-
-  for (;;) {
-    if (Date.now() >= deadline) {
-      throw new Error('Timed out waiting for the Backstage ingest job to complete.');
-    }
-
-    let response: Response;
-    try {
-      response = await fetch(jobUrl, {
-        headers: { Authorization: `Bearer ${uploadToken}` },
-        signal: AbortSignal.timeout(BACKSTAGE_INGEST_POLL_REQUEST_TIMEOUT_MS),
-      });
-    } catch {
-      await Bun.sleep(BACKSTAGE_INGEST_POLL_INTERVAL_MS);
-      continue;
-    }
-    if (!response.ok) {
-      throw new Error(
-        `Backstage ingest job polling failed (${response.status} ${response.statusText}).`
-      );
-    }
-
-    const job = (await response.json()) as BackstageIngestJobResponse;
-    if (job.state === 'completed') {
-      if (typeof job.result !== 'string' || !job.result.trim()) {
-        throw new Error('Completed Backstage ingest job did not return a signed result.');
-      }
-      return job.result;
-    }
-    if (job.state === 'failed') {
-      throw new Error(`Backstage ingest job failed: ${job.reason}`);
-    }
-
-    await Bun.sleep(BACKSTAGE_INGEST_POLL_INTERVAL_MS);
-  }
-}
-
 async function uploadSourceResumably(
   authorization: UploadAuthorization,
   source: SourceArtifact
@@ -408,7 +362,9 @@ async function uploadSourceResumably(
           rejectUpload(error);
           return;
         }
-        void pollBackstageIngestJob(jobUrl, authorization.uploadToken).then(
+        void pollBackstageIngestJob(jobUrl, authorization.uploadToken, {
+          errorMessageTerminator: '.',
+        }).then(
           resolveUpload,
           rejectUpload
         );
